Update assignment.md
Aside from typo correction, I deducted the requirement of # loops in the last question, because I think this could take even more time for students with less optimal laptop hardware. I believe running fewer rounds would not hurt the conclusion and the insights too much, and make it more convenient for the students.